Fixed compilation errors - now compiles!
